issue#11 add sharedTest to use same mocks in test and androidTest
