--- mockinizer/src/androidTest/java/com/appham/mockinizer/Mocks.kt
@@ -1,9 +1,0 @@
-package com.appham.mockinizer
-
-import okhttp3.mockwebserver.MockResponse
-
-val mocks: Map<RequestFilter, MockResponse> = mapOf(
-    RequestFilter(path = "/typicode/demo/mocked") to MockResponse().apply {
-        setResponseCode(200)
-    }
-)
