+++ mockinizer/src/test/java/com/appham/mockinizer/MockinizerInterceptorTest.kt
@@ -1,7 +1,6 @@
 package com.appham.mockinizer
 
 import com.nhaarman.mockitokotlin2.*
-import okhttp3.Headers
 import okhttp3.Interceptor
 import okhttp3.Request
 import okhttp3.RequestBody.Companion.toRequestBody
@@ -14,34 +13,6 @@ import org.junit.jupiter.params.provider.MethodSource
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 internal class MockinizerInterceptorTest {
-
-    private val mocks: Map<RequestFilter, MockResponse> = mapOf(
-        RequestFilter(path = "/typicode/demo/mocked") to MockResponse().apply {
-            setResponseCode(200)
-        },
-        RequestFilter(path = "/typicode/demo/foo") to MockResponse().apply {
-            setResponseCode(200)
-        },
-        RequestFilter(path = "/typicode/demo/error500") to MockResponse().apply {
-            setResponseCode(500)
-        },
-        RequestFilter(path = "/typicode/demo/private") to MockResponse().apply {
-            setResponseCode(403)
-        },
-        RequestFilter(path = "/typicode/demo/delete", method = Method.DELETE ) to MockResponse().apply {
-            setResponseCode(200)
-        },
-        RequestFilter(path = "/typicode/demo/post", method = Method.POST, body = """{"hey":"ya"}""" ) to MockResponse().apply {
-            setResponseCode(200)
-            setBody("""{"foo":"bar"}""")
-        },
-        RequestFilter(path = "/typicode/demo/header", headers = Headers.headersOf("name", "value")) to MockResponse().apply {
-            setResponseCode(200)
-        },
-        RequestFilter(path = "/typicode/demo/headers", headers = Headers.headersOf("name", "value", "foo", "bar")) to MockResponse().apply {
-            setResponseCode(200)
-        }
-    )
 
     private val chain: Interceptor.Chain = mock()
